Ticket(s): #4836 - New Server Rule Apply Time Fix
Schema/Configuration Changes: N/A

Fixed issue with how we were getting the selected value after the rule type was changed.  We had to go to the Request information instead of using the SelectedIndex.  Also in the course of the fix, changed the list of rule types so they're sorted alphabetically.