Ticket(s): #4837 - dcm_tmp files being processed by reinventory
Schema/Configuration Changes: N/A

Not sure if this will fix the problem, but it appears its related to the DirectoryInfo.GetFiles() method searching for *.dcm files and getting back these plus *.dcm_tmp files.  Changing the code to do a second check to make sure the files are *.dcm files before processing.  Also made a change to how the FIlesystemStudyProcess works to take this into account, also.