Ticket(s): #2465 - Cleanup of Dispose/Abort/Close for Scu/Scp
Schema/Configuration Changes: N/A

Partially from patch submitted on the forums.  Reworked the Dispose of connections.  Introduces an Abort() method to NetworkBase and to ScuBase.  The Abort method can give a timeout as to how long to wait, or it can wait indefinitely for the Abort to complete()  the abort will send an Abort PDU if applicable, then wait 2.5 seconds for shutdown, and then force a shutdown if it doesn't happen naturally.  As part of this, decided to remove the Close() method from DicomClient.  The description of the method really described the new Abort() method, and thought it was redundant.

[t:2465]