Ticket(s): #3737 - Series Level Move updates
Schema/Configuration Changes: N/A

Updates in how series level moves are inserted into the database.  They are now put in the WorkQueueUid table instead of in the xml column of the WorkQueue table.  Also, updated the service for handling this.  There still a bug that the device is not passed properly to the code.

[t:3737]