Ticket(s): #3822 - Applying rules after edit
Schema/Configuration Changes: Added new ServerRuleApplyTimeEnum

Added support for applying rules as part of the edit process.  As part of this, a new ServerRuleApplyTimeEnum was added for SopEdited.  Also refactored the edit a bit and moved its functionality into the ClearCanvas.ImageServer.Core project.

Note that these changes will affect the plugins for editing for JDMI.  Will need to update them to reflect this.

[t:3822]