Ticket(s): #4487 - ReconcileStudy Refactoring
Schema/Configuration Changes: N/A

Fixed bug in prior checkin.  We had an issue that the Locking mechanism didn't work before.  I refactored and moved it, but it turns out it wasn't needed, because the study was already locked when the entered/exited from the SIQ.  Just commented out the code to get reconcile working again.

[t:4487]