Ticket(s): #4487 - Integrate StudyProcess in with ReconcileStudy WorkQueue entry
Schema/Configuration Changes: N/A

Changes to reconcilliation to to process the study right away.  Changed code to use the SopInstanceProcessor directly when the reconcilliation is done. Also made changes to the ServerCommandProcessor.  Added new IAggregateServerCommand interface tor defining commands that are aggregates, or have sub-commands.  These should now be rolled back properly.