Ticket: #5129 - Perform conversion of VS2005 solutions and projects to VS 2008
Fixed problem with how plugins were setup with the web app.  Changed it so there was no longer a plugins directory, and made proper adjustments in the build script to make this work.  Note that changes will be required for Jaguar to put the rules plugins in the proper place during the build.