Ticket: #5155 Redesign PD dialog so that it can be templated
Modified the mechanism for adding recipients so that there is now only a single lookup control, which allows finding both staff and staff groups.  This relies on a new class that was introduced called LookupHandlerAggregator, which is essentially an implementation of ILookupHandler that wraps any number of child lookup hanlders and aggregates the results.