Ticket: #5182 Improve user experience of adding recipients in PD dialog
Added machinery to the ITable/TableView framework to allow table columns to specify custom in-place editors.  This works by having the ITableColumn.GetCellEditor() method return an instance of an ITableCellEditor.  The class the implements ITableCellEditor must have an AssociateView attribute specifying a view for the table editor.  The view will be displayed in-place in the table.

This is not quite working perfectly yet.

Also did some code clean-up in TableView and SuggestComboBox classes.