Ticket: #3249 - Add StyleSheet options for Validation
Modified the Validator to be configurable using an CSS class for indicating invalid input. All validated inputs are affected, but no visual changes have been made, they all still use the Red border/yellow background, but this is specified using a class instead of specifying the color directly in the control.