Ticket: #5205 Need a way to opt out of PD review in one specific use case
Schema/Configuration Changes: N/A

In the case where there is no existing PD conversation, but the rad is verifying an ER procedure, a message box now pops up asking if the user would like to create a PD review prior to verifying.  If the user says 'yes', the PD dialog is shown.  If the user says 'no', the PD dialog is not shown and the verification proceeds as usual.