Ticket: #4878 Missing audit information in delete operation
Schema/Configuration Changes: N/A

The "old-value" of all non-collection properties is now audited for deleted entities.  Collection properties cannot be audited when the parent item is deleted because it causes NHibernate to throw exceptions.  Although we could theoretically try to dig down and figure out what the issue is, we don't really need to audit collection properties for deleted entities, as they don't contain any identifying information.