Ticket: #4866 In downtime mode, unable to search for a performed order in Performing FS
Schema/Configuration Changes: N/A

Did some refactoring that is needed to fix this ticket.  Basically, we needed to make the GetWorkItemKey() method a little more lenient - it now takes a param of type object, so that any worklist item can be passed.  It is up to the subclass to decide which classes of worklist item it will accept.