Ticket(s): #4693 - Study Instance UIDs with trailing period not dealt with properly
Schema/Configuration Changes: N/A

Started trying to address the problem, but I ran into a potential issue that if an image is sent with the UID with the trailing padding, then another image is sent with the UID without the trailing padding, the images end up in the same study directory, but with different study storage locations.  The solution to catch this when importing was too intrusive for the number of existing studies at UHN with the problem.  Decided in the end to just reject images with a study instance UID that has a trailing padding character.

Note, however, that this commit does have changes to make rebuild xml/reinventory work with these types of studies.  A regression of this functionality shoudl be done.