Ticket(s): #5036 - Add Index to DeviceGUID
Schema/Configuration Changes: Yes

Did a check on the production system, and the queries done when checking for the count of auto-routes to a device were doing an index scan on the ScheduledTime/WorkQueueStatusEnum/WorkQueueTypeEnum index.  Decided that the index on DeviceGUID would help with this query, potentially.  Also noticed a number of deadlocks in the logs from Auto-Route requests.  I suspect this may be caused by this query, and the index may alleviate the problem.

Also, during integration of the indexes, decided to remove the index on WorkQueuePriorityEnum.  This value is an included column in the main ScheduledTime index.  Suspect that the index isn't even being used for our queries.

Also, decided to add an additional index on the Device table on the AeTiltle/ServerPartitionGUID columns, which may help in this area also, plus it may help with association negotiation times, since a query is done against the AeTitle for this.

Finally, noticed thta the GroupID index on the workqueue was associated with the incorrect database file group.  Moved this to the INDEXES filegroup.