Ticket: #5415 - Code Review Updates
Schema/Configuration Changes: Yes, renamed QueueData column in StudyIntegrityQueue to Details

Renamed the QueueData column in the StudyIntegrityQueue.  As part of the change, changed the model generator to now use auto-properties for the properties in the model files.  Required a few more updates to references of the properties to clean up.