Ticket: #5106 - Study Search List Bug when Viewing/Moving on last page
Updated the GridViewPanel to do a databind, if the gridview is not already databound. This keeps the GridView from losing a Row. In the GridView, I removed a ClearSelections() call from the SaveState, so that when selecting a row, it remains selected unless it is deselected or the page is refreshed.
Updated the EntityBroker to correct a bug where the first page of a search result returned 24 instead of 25 rows.