Ticket: #5134 - Remove HasUnprocessedCopy
Schema/Configuration Changes: N/A

After some review, decided to refactor the old code to increase performance.  The previous code found the location on disk of an SIQ and WQ entry, then went to look for the actual file's existence on disk.  This wasn't really necessary, because the existence of an unprocessed SOP is really stored in the WorkQueueUid and StudyIntegrityUid tables.  Made some changes so that a simple count(*) query using the entity brokers with a related condition to the Uid tables could determine if there was an unprocessed copy.  Did some profiling, and although there's still a minor delay for this operation, its quite a bit faster than the old method of checking the filesystem.  This should increase performance of dicom transfers in.