Ticket: #5479 - Orphan directories in Reconcile directory
Schema/Configuration Changes: N/A

Added an utility to scan the reconcile folder for potential problems. Folders are categorized as "empty", "contains only backup/temp files", "deleted afterward", "resent afterward" (storage record was inserted after the folder had been created, which suggested the study was removed from the database and resent).

The utility is put in the TestApp project for the time being.