Ticket: #5788 - Cleanup some innefficiences in StudyController
Schema/Configuration Changes: N/A

Cleaned up some cases where we had the StudyStorageKey in the code in a Study object, but we still tried to look up the value in the database using the ServerPartitionKey. This was still a leftover from the 1.3 release where we didn't have StudyStorageKey. This was still a leftover from the 1.3 release where we didn't have StudyStorage