Ticket: #5811 - Need method to only attempt to migrate lossless studies
Schema/Configuration Changes: N/A

Implemented support for migrating only lossless studies.  There's now a checkbox on the Migration Setup dialog that tells if only lossless studies should be attempted to be migrated.  We also added a new kind of QueryQueueTypeEnum, which signals a Lossles study migration.  Had to expand the plugin mechanism for archive customization to include a new call to check if a study is lossless.