Ticket: #5770 - Include query parameters in log file
Schema/Configuration Changes: N/A

Upon research of this ticket, discovered that the Query audit message was not implemented correctly according to DICOM Supplement 95.  A base64 encoded version of the DICOM query message was supposed to be included in the log message.  Implemented this feature for the ImageServer and ImageViewer.  Note that we may still add a human readable form of the query data in this log message, just to ease searching through the audit log.