Ticket: #5853 - Web GUI query is not audit logged Ticket: #5770 - Include query parameters in log file
Schema/Configuration Changes: N/A

Added audit logging of the Query message.  Also, cleaned up a bit more the Query audit log for dicom queries.  The transfer syntax of the base64 encoded message was not set properly, it was supposed to be a TypeValuePair.  Also added the logging of text based query parameters into the ParticipantObjectDetailString field of the audit message.