Ticket: #5866 - SIQ entries failing in the WorkQueue when the Patients Birth Date is invalid
Schema/Configuration Changes: N/A

Made updates to fix this problem.  When a value is invalid, we go back to what the prior value was and set/leave that value.  We also now catch any exception that may happen when encoding tags.  A warning message is logged when this happens.  Also made some changes to the TestApp application when debugging this ticket.