Ticket: #5297  - Users should not be allowed to reconcile study that is lossy compressed but has been archived as lossless
Schema/Configuration Changes: N/A

Part of this fix is made for ticket #6016 to change how the cache behaves. Assuming objects are only held in the cache for 30 seconds, the importer will pause for 60 seconds if it detects at least one of
the studies in the Incoming folder needs to be restored first. This is not the best solution because ideally we should be able to process images for studies that don't need to be restored first but that will increase the complexity of the code.