Ticket: #5495  - Deleting a study may result in orphaned reconcile folders
Schema/Configuration Changes: N/A

Fixed problem when deleting reconcile folder for duplicates. The path was not saved in the SIQ entry xml.

NOTE: the path is saved in the xml but is no longer used to find the folder. ServerPlatform.GetDuplicateGroupPath() is now used instead, which will resolve the path based on the latest filesystem path.