Ticket: #5989 - Some studies cannot be edited
Schema/Configuration Changes: N/A

Fixed problems with editing studies with the 'box' character.  Made two changes, one made update to how the study history is done so that we remove the character when placing the history record in the database.  Secondly, made a change where these characters are now stripped out of incoming images in the dicom service.  So we shouldn't have any more studies like thsi in the system going forward.