Ticket: #5106 - Study Search List Bug when Viewing/Moving on last page
Schema/Configuration Changes: N/A

Rolled back fix for this ticket.  Discovered that the fix was causing extra data binds within the web GUI, which would slow down performance of the GUI on large databases.  The fix for the ticket is not worth the performance drop.  Will defer this ticket in order to attempt to solve it better in the future.