Ticket: #5484  - Add DB Update Scripts for Performance milestone
Schema/Configuration Changes: Small upgrade script change

As per last comment on the ticket, the CannedText field didn't come through on a db upgrade.  Noticed one update statement in the upgrade script wasn't needed.  Removed this from the script, then manually ran an upgrade, starting with the 1.5 tables and then running the 2.0 upgrade script, and checked the CannedText table, and it was upgraded properly.