Ticket: #5304 - Could not change patient DOB
Schema/Configuration Changes: Updated web.config

Changed the format used for date input. It's now fixed to "dd?MM?yyyy", where ? is the date separator for to the current UICulture. For eg, if uiCulture in <globalization> of the Web.config is "es-CL", the date format will be dd-MM-yyyy. The format is applied to both the patient's birthdate and study date fields in the edit dialog.

InputDateFormat in the Web.config has been removed.