Ticket: #6083 - Page navigation on the alerts page is broken
Schema/Configuration Changes: N/A

The problem was caused by the update panel not being properly refreshed when the grid was databound. Also added some code to ensure the grid is refreshed when alerts are removed.

Fixed another problem where the last row in previous page is repeated on the next page. This fix should be reflected in all pages.