Ticket: #6101 - Deleting a series from a lossy compressed study causes the study to be re-archived as lossy
Schema/Configuration Changes: N/A

The flag for the javascript that enables/disables the series delete button was only doing a simple check if there were work queue items for a series before it would be deleted, instead of doing a comprehensive check to make sure the study was lossy, etc.  Updated the code to take this into account.  Made changes to block out the button, and also a double check before committing incase the state changes after the page has been rendered to the client.